homework/requirements.txt
django==3.2
redis==4.3.4
djangorestframework==3.14.0
oauthlib==3.1.0
oauth2client==4.1.3
django-oauth-toolkit==1.3.2
django-cors-headers==3.13.0
django-filter==22.1
django-extensions==3.2.1
django-redis==5.2.0
ipython==7.34.0